Store is_executable inside blob objects themselves
The existing code could not handle hard-linking of executable files and
had to copy them from `objects` to the destination path and toggle the
executable bit every time. This change enables executable blobs to be
stored separately in the object store, allowing them to be hard-linked
just like regular files for more efficient deduplication.

However, it means that two blob objects with identical byte contents but
different permissions will have to be stored separately. There is no way
in a Unix-like operating system to hard-link a file to a new location
and set only one of them as executable, so these two otherwise identical
files will have to be fetched separately and stored separately on disk.

Whether this trade-off is worth it in the long run isn't immediately
clear. If binaries contain absolute path references to other checked-out
files in the store, there will only ever be a single copy of each binary
file present in the store at the time, making this change redundant. On
the other hand, if binaries have paths stripped out of them (path
rewriting for relocatable store binaries and/or for reducing redundant
downloads), then hard-linking those files is an impossibility anyway.