Add ObjectKind enum and return it in Objects iterator
Opening and buffering each and every store object into memory and
returning them in the `Objects` iterator is a terrible idea. We should
instead return an iterator of `(ObjectId, ObjectKind)`, which is much
more lightweight. Any objects the caller is specifically interested can
be fetched using `get_object()` trait method.

A positive side effect of this change is that `ObjectKind` centralizes
the logic for handling object file extensions, i.e. *.{blob,tree,pkg},
all in one place. Also, adding this as an optional argument to
`get_object()` and `contains_object()` should also make random access of
objects in the `FsStore` much faster, but this will have to come in a
separate commit.