Add optional ObjectKind arg to {get,contains}_object()
Specifying the `ObjectKind` upfront should drastically improve random
access performance for locating objects in a filesystem-based store,
which relies on file extensions to store the object types on disk. This
information is usually known ahead of time, so the fallback should
rarely be used in practice.

This argument is totally ignored by the in-memory store implementation.
It is only recommended for use with the filesystem-backed store to
improve performance, as mentioned before.