Implement compute_closure() helper function
This function traverses a DAG from the given roots all the way down to
the leaves, returning the closure topologically sorted such that if item
P depends on Q, then P is ordered before Q. This important property
makes delta computation very efficient when comparing the contents of
two stores.

Let's imagine a scenario where the user attempts to copy a closure of
one or more packages to a remote store. The first step is to establish a
connection to the remote store and send over a closure containing
objects that the client would like to copy. We need to determine which
store objects the server already has in order to avoid sending them over
the network redundantly.

To do this, we send our topologically-sorted closure entry-by-entry, and
the server ACKs each one it already has locally. As soon as it hits just
one object that it lacks locally, we know that every other object ID
after this point must be missing as well and needs to be copied over.
These remaining objects are fetched from the store, serialized into a
compressed stream of bytes, ideally chunked using something like FastCDC
for efficient storage and transfer, and deserialized to disk on the
other side.

Note that this helper function is only an implementation detail. The
public `Store::closure_for()` provided trait method will be implemented
in a separate commit.