Add init and bare_init FsStore initialization modes
Note that this changes the semantics of `FsStore::open()` to fail if the
path does not point to a valid store directory. This behavior has been
supplanted by `FsStore::init()`, which will open the given store
directory if it already exists or otherwise silently initialize one at
the given path (creating a new directory if not present). By contrast,
the stricter `FsStore::init_bare()` will initialize a new store _only_
if the target path exists, is a directory, and is empty.