Use dynamically sized Vec<u8> for Storage::Inline
It is actually a bug to use a zeroed `Box<[u8]>` initialized with a
constant length in `PagedBuffer`, and we shouldn't do it. This is for
two reasons:

1. It's very inefficient to allocate the full size in-memory buffer up
   front like that, considering that some files are relatively small and
   may not take up much space. We shouldn't needlessly waste space.
2. It's semantically wrong. While the `std::io::Write` implementation is
   able to correctly detect and handle the spillover from `Inline` to
   `File`, the `std::io::Read` implementation cannot. Because the length
   of the inline buffer is always == `threshold`, readers could
   potentially read past the end of the data and find a stream of `0u8`
   until EOF. Whoops!

This commit fixes these issues by replacing `Box<[u8]>` with `Vec<u8>`
as the in-memory buffer of choice, which we probably should've gone with
in the first place. Although we don't currently know what a good
up-front capacity for this `Vec` should be, it should at least fix any
potential bugs with the `std::io::Read` implementation.