Calculate tree/package hashes with JSON instead of fnv hash
It has come to my attention that the behavior of `#[derive(Hash)]` in
Rust is explicitly marked as not stable between language versions,
making the hashes for `Tree` and `Package` objects not stable as well.
To avoid this, we drop the use of `std::hash::Hash` and
`#[derive(Hash)]` and just feed the `id::Hasher` with the output of
`serde_json::to_vec` instead. This is much slower than the current
approach, sadly, but at least the output is guaranteed to be stable.