Add faster custom implementation of std::io::copy()
This function uses a 64KB buffer instead of the 8KB buffer used in
`std::io::copy()`, and it delivers measurably better throughput in most
cases. In my own benchmarks, `Blob::from_path()` was previously able to
process a 1GB file in around 1950-2100 ms, and now it can process the
same file in around 1500-1550 ms.

Adapted from this original source:

https://github.com/BLAKE3-team/BLAKE3/blob/e0f193ddc9c0400262dae4118b0660ae8d70586e/b3sum/src/main.rs#L256-L274

Note that this is far from the biggest bottleneck in the system, though.
The biggest overhead actually comes from `file.seek(SeekFrom::Start(0))`
which, if removed, cranks out crazy numbers like 550 ms. Unfortunately,
we need to have this seek in place because we usually would like to read
the file again immediately afterwards. Our current design requires us to
basically read the file in its entirety twice: once to hash the file in
`Blob::from_path()` and again in `Store::insert_object()` with a
`std::io::copy()` call.

A possible solution to this dilemma could be to maybe back `Blob` with a
`PagedBuffer` by default, and add a `.persist()` method which will do
either of the following:

1. Create a file at the given path, copy the data to disk, and normalize
   the permissions, if the `PagedBuffer` is inline.
2. Normalize the permissions of the temporary file backing the
   `PagedBuffer` and move it to the given path using `std::fs::rename()`
   which is much faster than doing I/O.