Write blob to tempfile instead of sink()/seek()
The use of a temporary file prevents a situation where the contents of
the file change after the `Blob` has already been created, leaving the
object ID out of date.

Benchmarks with `std::time::Instant` on my system show comparable times
for hashing a 1GB file of around 1450-1650 ms between hashing + draining
into a tempfile vs. hashing + reusing the old file with a `.seek()`
call to move the cursor back to the begining. With enough repetitions, I
saw some even lower timings around 1390-1400 ms, but this is likely just
the result of stuff being cached.