Refactor Store as a struct, add Remote trait, impl copy_closure()
This large refactor changes the design approach quite a bit. We now
represent `Store` as a struct containing a `B: Backend` instead of being
a pure trait on its own. The reasons for this are twofold: to reduce
duplication of logic between implementations (now we can have data
common between implementations, not methods alone), and also to simplify
the trait design, esp. with closure copying.

In regards to the first point, the biggest issue with having `Store` as
a trait was forthcoming, which is how to handle configuration and
building. The future implementation would have to be duplicated between
both `FsStore` and `MemoryStore`, and while shared structs would help
with this, it would be very difficult to maintain. Changing `Store` into
a struct fits the bill much more nicely and allows us to share
configuration between implementations.

As for the second point, the proposed design for closure copying was
very clunky under the store-as-trait design. Proposed designs
necessitated `self.create_pack()` and `dest.receive_pack()` methods
which had to buffer to a file or an in-memory buffer in between, which
was really not ideal. The other issue was that certain operations didn't
make sense on remote stores vs. local stores, e.g. `iter_objects()` and
`serve()` for remotes and `ping()` for local stores.

It only made sense to separate the `Store` trait into local and remote
variants, and now the packfile I/O can be nicely abstracted away behind
the `Remote` trait without polluting the `Store` logic. We can also copy
closures directly from any filesystem/in-memory store to another without
needing to compress and extract a pack in between. The use of packfiles
can now be reserved only for remote sources instead of local ones.