Write temp files to /var/tmp
This is because `std::fs::temp_dir()` on Linux refers to `/tmp` on most
systems, which is a `tmpfs` that isn't backed by the physical disk. This
means that the `persist()` method always fails, since it cannot be
atomically renamed to the final location. As such, we should write to
`/var/tmp` instead, which (according to the FHS) is intended to persist
across reboots and must therefore be backed by the local disk.

This globally writable directory should be available on both macOS and
Linux, so it seems like a natural choice for preserving the atomic
guarantees of `NamedTempFile::persist()`.

There is also a single use of `tempfile::tempdir()` which we update to
`tempfile::tempdir_in()` for similar reasons.