Improve PagedBuffer, rename to SpooledTempFile
I failed to notice that `tempfile::SpooledTempFile` already exists and
does the same thing as `PagedBuffer`, except it doesn't provide any
mechanism to persist the temp file to the store. This commit renames
`PagedBuffer` to `SpooledTempFile` and also shamelessly incorporates a
few small tweaks to improve efficiency and correctness that were
previously overlooked, including updating the seek cursor upon
overflowing to to a file.