Reintroduce lighter weight InstallName struct
We might actually need more than a `String` to represent the directory
name of installed packages, because we may want to retrieve the name and
`ObjectId` components from it without manual string parsing.

To remedy this, while keeping the `InstallName` type as lightweight as
possible, we simply create a newtype around `String` which does cheap
string slice parsing, and we also implement `AsRef<Path>` to avoid the
overhead of `Display` or `ToString`.