Ensure 'objects/##' subdirectory is created atomically
The current implementation of `Filesystem::insert_object()` is not a
completely atomic transaction. If a new empty `objects/##` subdirectory
is created, and the object we try to persist encounters an error, the
store is left with an empty `objects/##` subdirectory.

To resolve this, we must either:

A. Persist directly to `objects/##/<id>.<kind>`, if `objects/##` exists.
B. Persist to `<temp>/<id>.<kind>` and rename `<temp>` to `object/##`.
   If the `objects/##` directory is created concurrently, we atomically
   move the `<id>.<kind>` file into the `objects/##/` directory.

This ensures `insert_object()` behaves as a truly atomic transaction.