Refactor Blob internal representation, remove Read impl
The current `Blob` internal representation bends over backwards to cater
to the `std::io::Read` implementation, even though this operation is
very rarely used in practice, and the code for it is pretty gross.

1. Blobs created via `{from_path,from_reader,from_path_unchecked}`
   always keep their file descriptors and/or memory maps open, even when
   it's totally unnecessary. We repeatedly open and close FDs/mmaps each
   time a `Blob` object is created, retrieved, inserted, or dropped. As
   you might imagine, this can be needlessly expensive. The reason why
   we must keep the files open is because the `std::io::Read` impl could
   be called at any moment.
2. `Blob::{from_reader,from_path}` both leave the I/O cursor at the end
   of the file, rather than at the beginning, because seeking back to
   the beginning of a possibly large file is needlessly expensive. This
   means that `std::io::Read` does't even behave properly for `Blobs`
   created from those constructors. Why do we implement `std::io::Read`
   again?
3. The proliferation of `pub(crate)` constructors and `Kind` variants
   indicates that the abstraction is very leaky.
4. There is significant overlap between the `object` and `install`
   modules. It would be nice if hashing files modulo some text could be
   moved out of `install` into a new constructor instead, e.g. something
   like `Blob::from_reader_rewrite()`.

This commit attempts to address all of these concerns. In short, we make
the internal representation of `Blob` much lighter, remove the direct
`std::io::Read` implementation, and provide an explicit
`into_contents()` method for when the caller actually wants to `Read`
the file contents. This deferred reading approach makes `Blob` objects
extremely cheap to instantiate and pass around. In fact, retrieving a
`Blob` from a store is now basically zero-cost until the user calls
`into_content()` to start reading the file content.

Additionally, we add a new `Blob::from_reader_rewrite()` constructor and
shuffled the module structure around a bit. Notably, `reference.rs` is
now an implementation detail of `Blob` and is fully abstracted away.