Replace Blob::from_reader() with Blob::from_writer()
This `Blob::from_writer` interface is significantly more flexible
because it lets the caller decide on the buffering strategy, if any, and
possibly wrap it in a `RewriteSink` if necessary. It also cuts down on
the number of private and `pub(crate)` constructor variants that `Blob`
has, greatly simplifying the interface.

We also remove the unused Criterion benchmarks that still depend on it.