Split Remote trait, add copy_closure() free function
This commit attempts to improve upon the `Remote` trait, which
previously conflated the behaviors of remote sources and remote
destinations, by splitting its functionality between two distinct
`Source` and `Destination` traits.

We also turn `copy_closure()` into a free function which would allow us
to copy from any `Source` any `Destination`.

This new design is capable of supporting the following copy scenarios:

1. Local src  -> Local dst
2. Local src  -> Remote dst
3. Remote src -> Local dst
4. Remote src -> Remote dst

The introduction of the `Source` trait also relaxes the requirement that
sources be a `Store<B>`. Instead, one could potentially copy from a
binary cache to a store, or a store to a binary cache, or even between
two binary caches in the future.

See this StackExchange post for the reasoning behind `scp -3` that
explains why we route traffic between remote src -> remote dst through
the local machine in our implementation:

https://superuser.com/a/686527