Move Store into local.rs, move Objects into root
This is the beginning of an effort to transition `Store` functionality
into a trait, rather than a struct, and also expose `Objects` as a
top-level trait for use with store-like things that can't build packages
from source and only contain objects, e.g. binary caches, to be added in
the future.