Move install module into local module
Only local stores (filesystem or in-memory) will be able to serialize
packages from directories, which makes sense intuitively. Not even
daemon stores will need this functionality, since a daemon store is just
a dumb a UNIX socket connection to a local store, which does have this
functionality.

The overall module structure is still pretty messy, so the next few
commits will try to smooth things out a bit.