Add provided compute_closure() method to Objects trait
Considering local stores and binary caches will use likely use this
method, it's probably best that this be included on the `Objects` trait
as a provided method. For remote implementations, e.g. over SSH or gRPC
or HTTP or something, we can override this method to make a remote call
instead.