Add Store trait, rename Store struct to LocalStore
As explained in the previous few commits, `Store` will be turned into a
trait and implemented on both local and daemon based store
implementations. The `LocalStore` struct will still support two
backends, `Filesystem` and `InMemory`, later on.

The reason why `Store` can't feasibly stay a struct with `Filesystem`,
`InMemory`, and `Daemon` backends is because `Daemon` is a special case.
In this hypothetical `Daemon` implementation, the `Packages` and
`Objects` fields and the `install_package()` function would have to be
totally inaccessable by clients, with only the privileged daemon having
the permission to launch builds and serialize packages from directories.
As such, it means that this hypothetical `Daemon` store backend _must_
implement a high-level `Store` trait which just forwards directly to a
secured Unix domain socket.

And if `Daemon` gets this special treatment, then clearly there are two
categories of stores that exist: local stores (FS and in-memory) and
remote stores (daemon, possibly SSH?). Local stores can have `open()`,
`init()`, and `init_bare()` methods, while remote stores do not.