Check existence of Package object, not package install dir
We have previously established that while inserting into the `objects`
directory itself is an atomic operation and the instantiation of
packages in the `packages` dir is also atomic, the full process of
inserting a package object and then instantiating it is decidedly _not_
atomic because there are two steps involved.

In fact, it is the same situation with Nix (first, move the prepared
package dir into `/nix/store` and then register metadata in the DB). If
one of the requisite steps fails in the middle, then the package is
considered not to be installed at all. There might be a stale entry in
`packages` left behind and it will need to be cleaned away at a later
point by a GC run or something, but at no point have the invariants of
the Merkle tree been broken.

This means that the source of truth for whether a package is installed
or not is ultimately the presence of the `.pkg` file in the `objects`
directory. The `packages` directory, then, is simply a "view" into the
`objects` directory; it cannot be blindly relied upon when checking for
integrity. Therefore, when a package is being installed, we must check
for references by consulting `objects` and not the existence of an
install directory in `packages`.

Therefore, this commit gets rid of the `Packages::contains()` method for
now and changes the provided `Packages::install()` method to use the
`objects` store as the definitive source of truth. This change should
improve the integrity of the store overall in the face of I/O errors.