Remove progress stream from CopyInfo, rename to Summary
This data model simply cannot be made to work as advertised.
`insert_objects()` must block until the closure has been fully copied
over, but the progress stream must be returned immediately, somehow. One
could spawn a thread or two to keep things concurrent, but this is
wasteful. Perhaps it may be better to let `copy_closure()` and
`insert_objects()` block, but supply a progress callback instead. This
will be added in a later commit.