Make pack reading/writing async, make copy_closure() async
There is a well-documented impedance mismatch between the `Source` and
`Destination` traits we have today, and the need for reporting
byte-level progress from `copy_closure()`. This is because those two
traits only operate on the level of whole store objects, which is much
too granular for our needs.

This commit integrates the `PackWriter` and reader into these traits,
which were unused until now, in order to serialize the stream of objects
to bytes and back again during the copy, so we can use it as a source
for measuring progress. The actual progress reporting will come in a
separate commit.

Unfortunately, the data model for `copy_closure()` had to be tweaked as
well, requiring two separate read/write tasks to be running in separate
threads, while the main thread listens to the progress channel and ticks
the progress bar. After surveying prior art, experimenting with using
scoped threads and channels to pass progress notifications and propagate
errors from both the reader and writer threads, I don't think there is a
sufficiently reliable and fully deterministic way to halt the copy
part-way and propagate the _most meaningful_ error to the main thread.

So I decided to bite the bullet and partially introduce async/await
into the codebase. I tried to avoid it till now, due to the heavily
filesystem oriented design not being well-suited for async I/O and the
increase in complexity in other parts of the program, but ultimately,
the ability to join multiple futures concurrently on a single thread and
deterministically propagate errors back to the main thread seemed like a
perfect fit for what we want to do.

The silver lining is that we likely would've had to do this anyway,
since we plan to support remote stores in the future.